lint_requirements.txt
click == 7.1.2
colorama == 0.4.4
numpy == 1.17.3
opencv-contrib-python >= 4.5.2.54, != 4.5.4.58, <= 4.5.5.64
protobuf <= 3.20.1
pyyaml >= 5.3
requests == 2.24.0
tensorflow == 2.2.0
torch == 1.10.0
torchvision == 0.11.1
tqdm == 4.45.0
typeguard >= 2.13.3
scipy >= 1.4.1